added checks for unique ids in db and unique names in db

+++ backend/src/main/java/com/example/react_web_project_backend/repository/UserService.java
@@ -26,12 +26,17 @@ public class UserService {
     }
 
     public User addUser(String name, String password) {
+        if (userRepository.existsByName(name)) {
+            throw new IllegalArgumentException("User with name already exists");
+        }
+
         UUID id = UUID.randomUUID();
         User user = new User(id, name, password, new HashMap<>());
         return userRepository.save(user);
     }
 
     public boolean deleteUser(UUID id) {
+        if (!userRepository.existsById(id)) return false;
         return userRepository.deleteUser(id);
     }
 }
